agent/requirements.txt
# Customer Retention Agent Dependencies
strands-agents>=0.1.0
boto3>=1.34.0
botocore>=1.34.0
requests>=2.31.0
bedrock_agentcore>=0.1.7
